+++ src/data/Property.ts
@@ -1,0 +1,4 @@
+export type PropertyDocument = {
+    key: string;
+    value: string;
+}