+++ src/services/ImageMaker.ts
@@ -1,18 +1,17 @@
 import sharp from 'sharp';
 import path from 'path';
 import fs from 'fs';
-import { getVersion } from '../commands/slashCommands/version';
+import { version } from '../commands/slashCommands/version';
 
-const TextToSVG = require('text-to-svg');
-const textToSVG = TextToSVG.loadSync();
- 
-const attributes = {fill: "red", stroke: "black"};
-const options = {x: 0, y: 0, fontSize: 42, anchor: 'top', attributes: attributes};
+const textToSVG = require('text-to-svg').loadSync();
  
 
 export class ImageMaker {
     public static async getVersionImage(): Promise<Buffer> {
-        const topText = textToSVG.getSVG('Bella version ' + getVersion(), options);
+        const attributes = {fill: "red", stroke: "black"};
+        const options = {x: 0, y: 0, fontSize: 42, anchor: 'top', attributes: attributes};
+
+        const topText = textToSVG.getSVG('Bella version ' + version, options);
         const bottomText = textToSVG.getSVG("IS HERE", options);
         
 		return sharp(path.join(__dirname, "../../assets/power.jpg"))